+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Changelog
 Next
 ----
 
+- Breaking change: The required Sphinx version is at least 4.0.
+- ``sphinx-prompt`` is now an optional dependency.
+  Thanks go to @dgarcia360 for this change.
+
 2020.09.30.0
 ------------